requirements.txt
openai>=1.30.0
streamlit>=1.33.0
python-docx>=1.1.0
reportlab>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-dotenv>=1.0.0
